Lets user choose how to retrieve his highlights. (#29)
* Lets user choose how to retrieve his highlights.

The user can decide whether to fetch the content from a kindle device,
or log in to Amazon Cloud Reader. According to the user selection, the
items are rendered in the Home view.

* Update main.development.js

* Minor code changes.

* Minor code changes

* Update main.development.js

* Lets the user select the kindle device

* Solves a problem when the path was passed to KindleParser.

KindleParser assumes that the path is a string, but the dialog box returned
a Path object. Now, once the path has been returned from the dialog box, it
is casted as string and then passed to the parser.

* Sync annotations from device.

Lets the user select where are his/her annotations.

* Lets the user log out and transition to the login view.

* Add info message.

* Fixes login, implements fetch, changes dialog's icon.

- Solves a problem preventing users to sign in.
- Lets the user fetch new books/annotations according to their sync option.
- Changes dialog's icon to Kintrospect's logo.

* Fixes typo

* Changes the style of the select My Clippings dialog.